requirements.txt
opencv-python-headless==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
flask==2.3.3
python-dotenv==1.0.0
pyttsx3==2.90
Pillow==10.0.0
setuptools>=69.0.0
wheel>=0.42.0 